Split types from the declaration file
